psl-lint package configuration
