+++ src/language/codeQuality.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 import * as format from './pslFormat';
 import * as parser from '../parser/parser';
-import { getDiagnostics } from '../pslLint/main';
+import { getDiagnostics } from '../pslLint/activate';
 import { Diagnostic } from '../pslLint/api';
 import { PSL_MODE, BATCH_MODE, TRIG_MODE } from '../extension';
