+++ src/pslLint/cli/tsconfig.json
@@ -1,0 +1,15 @@
+{
+    "compilerOptions": {
+        "module": "commonjs",
+        "target": "es6",
+        "outDir": "lib",
+        "lib": [
+            "es6"
+        ],
+        "sourceMap": false,
+        "rootDir": "./../.."
+    },
+    "exclude": [
+        "node_modules"
+    ]
+}
